add keeper session manager
