+++ redis/redis-checker/src/main/java/com/ctrip/xpipe/redis/checker/healthcheck/session/KeeperInstanceManager.java
@@ -1,0 +1,10 @@
+package com.ctrip.xpipe.redis.checker.healthcheck.session;
+
+/**
+ * @author yu
+ * <p>
+ * 2023/10/31
+ */
+public interface KeeperInstanceManager extends InstanceSessionManager{
+
+}
