+++ redis/redis-checker/src/main/java/com/ctrip/xpipe/redis/checker/healthcheck/session/KeeperSessionManager.java
@@ -1,0 +1,9 @@
+package com.ctrip.xpipe.redis.checker.healthcheck.session;
+
+/**
+ * @author yu
+ * <p>
+ * 2023/10/31
+ */
+public interface KeeperSessionManager extends InstanceSessionManager{
+}
