+++ redis/redis-checker/src/main/java/com/ctrip/xpipe/redis/checker/healthcheck/session/RedisSessionManager.java
@@ -1,16 +1,10 @@
 package com.ctrip.xpipe.redis.checker.healthcheck.session;
 
-import com.ctrip.xpipe.api.endpoint.Endpoint;
-import com.ctrip.xpipe.endpoint.HostPort;
-
 /**
  * @author marsqing
  *
  * Dec 1, 2016 6:40:43 PM
  */
-public interface RedisSessionManager {
+public interface RedisSessionManager extends InstanceSessionManager {
 
-	RedisSession findOrCreateSession(Endpoint endpoint);
-
-	RedisSession findOrCreateSession(HostPort hostPort);
 }
