+++ redis/redis-checker/src/test/java/com/ctrip/xpipe/redis/checker/healthcheck/impl/DefaultHealthCheckerMockTest.java
@@ -2,14 +2,11 @@ package com.ctrip.xpipe.redis.checker.healthcheck.impl;
 
 import com.ctrip.xpipe.endpoint.HostPort;
 import com.ctrip.xpipe.redis.checker.AbstractCheckerTest;
+import com.ctrip.xpipe.redis.checker.CheckerConsoleService;
 import com.ctrip.xpipe.redis.checker.config.CheckerConfig;
 import com.ctrip.xpipe.redis.checker.healthcheck.HealthCheckInstanceManager;
 import com.ctrip.xpipe.redis.checker.healthcheck.meta.MetaChangeManager;
-import com.ctrip.xpipe.redis.core.entity.ClusterMeta;
-import com.ctrip.xpipe.redis.core.entity.DcMeta;
-import com.ctrip.xpipe.redis.core.entity.RedisMeta;
-import com.ctrip.xpipe.redis.core.entity.SourceMeta;
-import com.ctrip.xpipe.redis.core.entity.XpipeMeta;
+import com.ctrip.xpipe.redis.core.entity.*;
 import com.ctrip.xpipe.redis.core.meta.MetaCache;
 import com.google.common.collect.Sets;
 import org.junit.Assert;
@@ -20,7 +17,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
+import org.xml.sax.SAXException;
 
+import java.io.IOException;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
@@ -43,11 +42,16 @@ public class DefaultHealthCheckerMockTest extends AbstractCheckerTest {
     private MetaChangeManager metaChangeManager;
 
     @Mock
+    private CheckerConsoleService checkerConsoleService;
+
+    @Mock
     private CheckerConfig checkerConfig;
 
     @Before
-    public void setupDefaultHealthCheckerMockTest() {
+    public void setupDefaultHealthCheckerMockTest() throws IOException, SAXException {
         when(checkerConfig.getIgnoredHealthCheckDc()).thenReturn(Collections.emptySet());
+        when(checkerConfig.getConsoleAddress()).thenReturn("127.0.0.1");
+        when(checkerConsoleService.getXpipeAllDCMeta(Mockito.anyString(), Mockito.anyString())).thenReturn(getXpipeMeta());
         when(metaCache.getXpipeMeta()).thenReturn(getXpipeMeta());
     }
 
@@ -101,6 +105,7 @@ public class DefaultHealthCheckerMockTest extends AbstractCheckerTest {
         jqDcMeta.addCluster(clusterMeta);
         xpipeMeta.addDc(jqDcMeta);
         when(metaCache.getXpipeMeta()).thenReturn(xpipeMeta);
+        when(checkerConsoleService.getXpipeAllDCMeta(Mockito.anyString(), Mockito.anyString())).thenReturn(xpipeMeta);
 
         checker.generateHealthCheckInstances();
         verify(instanceManager, times(1)).getOrCreate(new ClusterMeta().setId("single_dc_cluster"));
